Add a "hail mary" strategy option for the final guess
The general idea for the strategies implemented, even the naive
strategy, is to narrow down the number of possible answers until only
1 remains, the correct one.

Unfortunately six guesses is not always enough to get there.
For that final guess, we're not longer narrowing, we just
hope that we can get lucky from the remaining guesses.

So it makes sense to use a different strategy, like "freq".